% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distsim.R
\name{multiplot}
\alias{multiplot}
\title{Arrange several plots into a single view}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{Set of plots to arrange.}

\item{plotlist}{List of plots}

\item{file}{Unused}

\item{cols}{Number of columns to arrange the plots into.}

\item{layout}{Layout matrix}
}
\description{
Arrange several plots into a single view
}
