% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distsim.R
\name{xbarthetadist}
\alias{xbarthetadist}
\title{Distribution of xbar and theta}
\usage{
xbarthetadist(n, iter, mu, sigma)
}
\arguments{
\item{n}{Number of samples per iteration}

\item{iter}{Number of iterations to simulate}

\item{mu}{Expected value of the variables}

\item{sigma}{Covariance of the variables}
}
\description{
Distribution of xbar and theta
}
